pyproject.toml
[project]
name = "spyrite"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pillow>=12.0.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.1",
]
